requirements.txt
imgviz==1.2.6
mahotas==1.4.13
numpy==1.21.4
opencv_python_headless==4.5.3.56
Pillow==9.2.0
PyQt5==5.15.7
PyYAML==6.0
